protocol2026/requirements.txt
appdirs==1.4.4
bcrypt==4.3.0
bokeh==3.7.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
DingtalkChatbot==1.5.7
docx2pdf==0.1.8
et_xmlfile==2.0.0
fonttools==4.56.0
future==1.0.0
greenlet==3.1.1
hexdump==3.3
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.6
scapy==2.7.0  # pip install scapy 默认安装最新可用版本；这里使用 == 2.7.0 锁定已验证可用版本
kiwisolver==1.4.8
ldap3==2.9.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
narwhals==1.30.0
netifaces==0.11.0
netmiko==4.5.0
ntc_templates==7.7.0
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
pillow==11.1.0
pinyin==0.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pygal==3.0.5
Pygments==2.19.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.2.1
pypinyin==0.53.0
pyserial==3.5
pyshark==0.6
pysnmp==7.1.16
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
resend==2.6.0
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
scp==0.15.0
six==1.17.0
SQLAlchemy==2.0.39
termcolor==2.5.0
textfsm==1.1.3
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
webencodings==0.5.1
xyzservices==2025.1.0
zipp==3.21.0
